Initial Commit
Domain Name System (DNS) is the internet backbone for providing a mapping between humanreadable
hostnames and computer understandable Internet Protocol (IP) addresses. It uses simple
UDP because of that, it contains some issues.
1. DNS tunneling is a method of using DNS protocol as a mean of encapsulating data transmission
between a client and a server.
2. This vulnerability makes DNS protocol highly susceptible to various active and passive attacks,
such as man-in-the-middle attacks (MitM) and eavesdropping.
In 2018, a new protocol named DoH was released, which not only enhances performance but also
improves user privacy and security by preventing eavesdropping and manipulation of DNS data
through MitM attacks. DoH wraps DNS records (both requests and responses) in an HTTPS
connection, providing encryption and authentication of the server, and changing the connectionless
aspect of DNS. DoH primarily serves two purposes–preventing on-path devices from interfering
with DNS operations and allowing web applications to access DNS information via existing browser
APIs.